requirements.txt
anyio==3.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-api-keys==0.5.1
google-cloud-core==2.3.2
google-cloud-dialogflow==2.15.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==1.0.0
python-telegram-bot==20.1
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
urllib3==1.26.14
vk-api==11.9.9
